Update README.md
This website serves as my personal portfolio, showcasing my accomplishments, abilities, and projects using HTML, CSS, and JavaScript. Visitors are given an engaging experience by its sleek, contemporary design, interactive features, and fluid animations. Because the website is fully responsive, it displays beautifully on desktop, tablet, and mobile devices.

The integrated contact form, which saves messages straight to Google Sheets, is a major feature of this portfolio. This facilitates the collection and systematic management of inquiries. In general, the website functions as a dynamic medium for presenting